support for undo zoom (zoom out)
Pressing Undo (CMD-z) will undo the previous "zoom in" step. There is
no limit on zoom-in and zoom-out(undo), but s of num zooming in beyond
a certain limit (1e-17) on X_Scale or Y_Scale will drop all resolution
because we are hitting the max resolution of "double" on Macbooks.